feat(owner): P0 fixes, moderation decision controls, profile sections
The owner's final verdict of 2026-09-16, worked in its stated order.

P0 — the twelve fixes
  * Login and Signup draw "Continue with Google", greyed and inert, over
    "Sign up / Continue with email". It navigates nowhere: the real OAuth
    hand-off is behind NEXT_PUBLIC_GOOGLE_AUTH and needs credentials the
    owner holds.
  * /profile gains Comments and Stats, switched by ?tab= so the three
    sections share one identity panel. Comments reads a new endpoint,
    GET /users/{id}/comments; Stats reads the trust profile, which now
    serves the next stage and what it needs — computed from the ladder in
    app/services/trust.py so the frontend carries no second copy of the
    thresholds. The frame's "Frequent Interactions" card is not drawn:
    there is no follow graph and no endpoint, so it would be three
    invented faces.
  * The account menu lists only real destinations, adds Moderate for a
    moderator or admin, and logs out through the server action.
  * Ratings run 0 to 5 in half steps (0048), and zero is an answer rather
    than a missing one.
  * Only the root owner may grant or revoke staff roles: the configured
    email AND the super-admin column, enforced server-side and audited.
    Checked against production first — the rule locks nobody out.
  * Admin user management shows handle and email, not opaque UUIDs.
  * One labelled QA seller exists in production, reversible with one
    DELETE, with no fabricated ratings, revenue or merchant identity.

Moderation decision controls (§30)
  The console could inspect and not answer. Now the queue's detail pane
  publishes, monetizes or rejects, and the Report tab dismisses, removes,
  restores or escalates — the last through a new endpoint and 0049, which
  gives a report an outcome so the queue stops re-offering work already
  done. Both panels are BUSINESS-REQUIRED / DESIGN-SYSTEM ALIGNED: frame
  6922:837 draws no controls, and they are built from this console's own
  panel and pill vocabulary rather than invented styling.

  The queue table is rebuilt to that frame's eight columns in its order.
  Where the frame draws a number this build cannot source, the cell still
  says so rather than showing a plausible one.

  components/moderation/{ModerationQueue,ReportQueue}.tsx are deleted:
  both were mounted on no route and called these endpoints through
  window.prompt, and a second implementation of a moderation decision is
  worse than none.

Docs and tests
  FULL_FEATURE_MATRIX rows 8.3, 8.6, 9.2, 9.4, 9.7, 9.8, 9.9 and X.3
  re-classified, X.12 – X.16 added, row count corrected to 119 with none
  unclassified. 3D/360 is DEFERRED — ASSET / CONTENT NOT AVAILABLE FOR
  THIS RELEASE, per the owner's clarification.

  Backend 1324 passed, 410 skipped (the skips are DB-gated and run in
  CI's isolated-database job). Frontend 285 passed — including the
  telemetry source test, which was failing on every Windows checkout over
  a carriage return and is now line-ending proof.

+++ backend/app/models/review.py
@@ -58,7 +58,8 @@ class Review(Base, UUIDPrimaryKey, Timestamps):
     verdict_explanation: Mapped[str | None] = mapped_column(Text)
     target_audience: Mapped[str | None] = mapped_column(Text)
     anti_target_audience: Mapped[str | None] = mapped_column(Text)
-    star_rating: Mapped[int] = mapped_column(SmallInteger, nullable=False)  # 1..5
+    #: 0 to 5 in half steps (0048). Zero is a real answer, not a missing one.
+    star_rating: Mapped[Decimal] = mapped_column(Numeric(2, 1), nullable=False)
     pros: Mapped[list | None] = mapped_column(JSONB)   # max 10 (app-validated)
     cons: Mapped[list | None] = mapped_column(JSONB)   # max 10 (app-validated)
